A couple of changes: 1.  When an error is detected, read the next line and append to the error message. 2. Also pick up miscellaneous warnings (warnings without a line number) and add them to the output.
With the changes I've made, it is possible to use latexmk.pl as the TM_LATEX_COMPILE command (in place of pdflatex).  There are a couple of bits of additional config that I'm working through to make it easy, but when I get it to work in a user friendly way I'll commit latexmk.pl to the Support directory (its gpl'ed) along with any changes to the generic build command.


git-svn-id: http://svn.textmate.org/trunk/Bundles/Latex.tmbundle@1396 dfb7d73b-c2ec-0310-8fea-fb051d288c6d