Update the LaTeX and view command so that we can use latexmk.pl.  TextMate still defaults to pdflatex but you can easily change that by setting TM_LATEX_COMPILER to latexmk.pl in the variables section of a project or in the preferences for TextMate.  Additional configuration of any personal options can be done in Support/lib/latexmkrc.
Since the output from multiple Latex runs all gets funneled to the same html window I have attempted to detect each run and reset the error and warning counters, so that the final lines of the html output would correctly reflect the error and warning counters for the final run.  However, the 'Run number' message is spit out by a Perl warn command which does not seem to be part of stdout or stderr.  Is there a way to tie whatever stream warn corresponds to with stdout?  I tried 2>&1 but that did not seem to work.

Brad



git-svn-id: http://svn.textmate.org/trunk/Bundles/Latex.tmbundle@1397 dfb7d73b-c2ec-0310-8fea-fb051d288c6d