Handle the log output more gracefully
Previously we would only check for the message about the written
transcript at the beginning of a log line. While this is usually the
case, there might be edge cases where the message actually follows
another message on the same line. The current fix for this bug was
proposed by xithan (https://github.com/xithan).