Don't follow “imports” inside generic macros
When we search for bibliography items or labels we also try to include
all files specified via `input` and `include`. We might however also
use `input` and `include` inside generic macros. This could can cause
trouble. For example: Before this commit we would to try to include the
“location” `Code/#1` when searching for cites if we used the following
TeX code:

    \newcommand{\codeinput}[1]
    {
        \begin{leftbar}
            \input{Code/#1}
        \end{leftbar}
    }