Simplify and modernize `run_viewer` function
We remove the option `force` and introduce
`suppress_pdf_output_textmate` instead. This new argument allows us to
remove the access to the global variables for the number of errors and
warnings inside `run_viewer`. We also add the argument
`tm_bundle_support`, which allows us to remove the access to the global
variable `TM_BUNDLE_SUPPORT`.

We also make the following changes to update the function according to
current Python coding practice:

- Add documentation and doc-tests

- Use lower case notation for function name and variables

 #ignore