Improve the `clean` command
We now use `find` to locate and delete the files. Before this change the
shell command would usually exit with a non zero code, since one of the
globs we used to match the auxiliary files would fail.

We now also display which files were removed by the `clean` command.