Extend and modernize `LatexParser`
We handle warnings containing line information better. Before this
change the output often lacked the last part of the warning which
contained the line where the warning occurred.

We also removed the function `handleFileLineWarning` and the
corresponding regex. It seems that warnings of this form are very rare.
A quick search on the web returned no results at all.

We removed the regex `^Error: pdflatex` since a search on the web
returned no result for errors of this form.

We also make the following changes to update the class according to
current Python coding practice:

- Add documentation and doc-tests

- Use lower case notation for function name and variables