Remove command `builtin` from `texmate`
It seems that the command was removed from the preferences a while ago.
Using `latexmk` seems to be a better option in most cases anyway, since
`latexmk` only calls the tex commands it needs to translate a file.
The `builtin` command on the other hand always used the same sequence
of tex commands to process a file, regardless of its structure.