Use single variable for location of tex file
Before this change we called `run-viewer` using a separate variable for
the directory and the the name of the tex file. We do not use the
directory name without the filename anywhere in the function. This
means we can use a simple interface for the function where we specify
the location of the tex file using only one variable.