Add support for gutter marks to “Watch Document”
This currently works best with a single tex document. If there are
multiple documents and there is an error in one of the included
documents, then the gutter mark in the included document will not
vanish even after the error was fixed. This is the same behaviour as
the one in “Typeset & View (PDF)”. Currently you can either just ignore
these gutter marks, restart TextMate or clean the gutter marks yourself
by using `mate -c error warning filename.tex`.