Change location of cache file for “Typeset & View (PDF)”
We now save cached typesetting data in the same directory as the
(master) tex file under the name `.filename.lb`, where `filename` is the
name of the tex file without its extension. This means that we need to
get the directory of the master file first, and therefore can not cache
this information anymore. This change will increase the average running
`texmate` a little. However, now we also should be able to share data
between runs in a more predictable manner.