Add parent of filepath to path in python scripts
In the moment this folder is always the directory `Support`. Adding
this path allows us to import functions from `lib`. Before this change
the folder was only added if `__package__` was unset. Imports from
`lib` must always work or running the scripts will fail. Therefore we
now always add `Support` to `sys.path`.

#ignore