Fix path of cache file in `texparser`
Before this commit it was possibly that the path of the cache file was
wrong, if the path of the tex file did not contain any information
about the directory. For example: If we called `texparser` with the
arguments `../Log/ünicöde.log ünicöde.tex`, then it would assume that
the cache file was located in `/.ünicöde.lb` instead of `./.ünicöde.lb`.

#ignore