Add folding markers for sections
You can now fold tex code like:

    \section{A section}

    Some text...

TextMate has — as far as I know — some limitations on the way it
handles foldings without end markers. Therefore the new folding markers
do not always work the way we want them to. For example: If you have
the following code

    \section{Section}

    Some text...

    \begin{equation}
        2 + 2 = 5
    \end{equation}

and fold on `\section`, then the folded section will not include the
environment `equation`. To include the environment in the folding you
can extend the code above so that it uses the old folding system:

    \section{Section} % (fold)

    Some text...

    \begin{equation}
        2 + 2 = 5
    \end{equation}

    % (end)