Use encoding `latin_1` to open log files
`latin_1` is probably *not* the correct encoding for tex log files. It
should however be much closer to the real encoding — probably `T1` —
than `utf-8`.

This commit should fix an error that occurred when `texmate` translated
files that used the package `inputencoding` together with the option
`utf-8`. The retranslation of extended characters such as “ü” caused
the log-file to contain characters that are not valid in the encoding
`utf-8`.

Thanks to all users, who reported the error. Especially Jan-Gerd
Tenberge who provided a minimal sample document.