Enhance “Open Master File”
If the current file is already the master file, then the command now
displays a tooltip informing us about that fact. This provides us with
feedback about the success of the command. Before that change we might
have thought that the command failed silently or was not executed at
all, when we invoked it inside the master file.