Add support for citation customization
You can now change the format of the snippet used in “Citation Based on
Current Word / Selection”.

For example: If you want to insert a reference of the form
`\cite[]{citekey}` where the caret is placed inside the square
brackets, then add the variable `TM_LATEX_CITE_SNIPPET` in
Preferences→Variables. Set the value of the variable to
`\cite[\$1]{CITEKEY}`.

You can change the format of the reference using snippet syntax. The
value of `CITEKEY` will be replaced with the citation key of the item
you selected. Do not forget to escape dollar signs inside the snippet!