Use case insensitive search in completion commands
We now use case insensitive regex search in “Label Based on Current
Word / Selection…”, “Citation Based on Current Word / Selection…” and
“Citation (Ref-TeX Style)”.

This means, that if the current word in your tex document is `bla`,
then invoking one of the aforementioned commands will not only search
for labels/citations containing the word `bla` but also for all case
insensitive versions of the word like e.g. `Bla` or `BlA`.

To *deactivate* this feature you need to set the variable
`TM_LATEX_SEARCH_CASE_SENSITIVE`:

1. Open Preferences→Variables
2. Add the following variable: TM_LATEX_SEARCH_CASE_SENSITIVE
3. Use whatever value you like. My recommendation would be to just
   leave the value field empty.
4. Select the checkbox on the left side of the variable