Improve heuristic for “Insert Citation/Label”
We now display a list containing all labels/citations if the last
character of the word is `{`, `}`, `~` or `,`. Before we only did this
if the current word was exactly one of those characters.

This means that we now get a full list of citations if we invoke
“Insert Citation” in the code `“Some Article”~‸`. Before we would —
with high probability — get no citations at all, because we only
displayed citations containing the current word `”~`.