Ignore commented code in “Insert Citation/Label”
The recent updates to the completion code in `latex` introduced a bug:
Even commented code was used as input source.

If your document contained the following code `%
\bibliography{ponies}`, then “Insert Citation” would try to read the
content of `ponies.bib`. This was especially problematic if the
specified file did not exist, since then the command would just display
a tooltip informing about the missing bibliography file.