Remove `YAML::Tiny` library from the bundle
We added `YAML::Tiny` when we tried to switch to `latexindent` as
default tool used by “Tidy”. Since we switched back to “LaTeX Tidy” we
do not need the library anymore. If you still want to use `latexindent`
to format your code, then you need to install `YAML::Tiny` manually.