Update “Help” document
We update the whole document with up-to-date information. The help text
is now a little bit easier to read, since we split long sentences into
smaller ones. The document also uses passive voice less often. If you
decides to read the help text, then at least in theory this update
reduces the risk of death by boredom slightly. We still do not
recommend to read the text in one sitting tough, since it is quite
long. Still, if you want to know more about the bundle, now might be a
good moment to open “Help”.