Update “Readme”
This commit adds a link to the help document to the “Readme”. This
should help people to get more familiar with the purpose of the code in
this repo. The “Readme” now also contains a little bit more information
about how to contribute. I also added a donation button, in the
unlikely case that someone wants to donate some money for my
contributions.

 #ignore