“Create Table”: Use default value for # of rows
The value for the number of rows is now optional. If you do not specify
the number of rows, then the command will just create a table with two
rows. E.g.: If you select the text `4` and invoke the command, then it
will create a table with two rows and four columns.