“Create Table”: Add new user interface
The old interface only showed a single line, where you could enter the
number of rows and columns. The new interface uses separate text fields
for the number of rows and columns. The new interface also adds a
checkbox “Tabular Only”. This checkbox determines if the command
creates a table environment or only a tabular environment.