Disable format commands inside scope `source`
Before this change it was possible to use the various format commands
such as “Bold”, “Italic” and “Underline” inside source code. For
example, if we select the code `print "Hi"` and hit `⌘` + `B` inside a
Literate Haskell script, then the command “Bold” would surround the
code with the command `\textbf`.