Fix Skim crash
“Typeset & View (PDF)” and “Jump to Current Line in Viewer” did not
work correctly if the user selected multiple lines. Instead of showing
the correct position in the PDF, the commands would just crash Skim.
This error was caused by `texmate` which called Skim's `displayline`
with an incorrect line number, such as `1-5`. The fixed version of
`texmate` now just sends the first line number of the selection to
`displayline`. For the previous example this is the number `1`.