Add new command “Convert Symbols in Word”
This command acts as replacement for the macro “Convert ASCII Symbol to
LaTeX Command”. “Convert Symbols in Word / Selection” replaces special
characters and certain sequences of characters with an equivalent LaTeX
symbol. It works either on the current selection, or – if there is no
selection — on the current word.

For example, if you select the following characters:

    <=> -> → =>

and press `⌘` + `&`, then TextMate replaces the selection with the
following LaTeX commands:

    \Leftrightarrow \rightarrow \rightarrow \Rightarrow