“Double Quotes”: Improve caret placement
The command “Wrap Selection in” > “Double Quotes” now places the caret
inside the double quotes (``'') if there is no selection. Before this
change it would place the caret after the quotes. This way you had to
move the caret two places to the left to insert text between the double
quotes, which was certainly not ideal.