“Clean”: Specify auxiliary files via regex
Before this change we specified auxiliary files via extension and
auxiliary directories via prefix. We now use a slightly extended
version of regular expressions for this task. The specification via
regular expression is more flexible, since it allows us to specify
directories and files that do not follow the schema used before.

 #ignore