“Clean”: Set `TM_BUNDLE_SUPPORT`
Before this commit we assumed that `TM_BUNDLE_SUPPORT` is always set to
the bundle support directory of the LaTeX bundle. This might not always
be the case though.

For example, if we check the doctests of `clean.rb`, with the command
“Run DocTest” of the “Ruby DocTest” bundle, then `TM_BUNDLE_SUPPORT` is
set to the bundle support directory of the “Ruby DocTest” bundle.
Consequently the doctests fail, since the path to the bundled `unf`
library  is incorrect.

 #ignore