Tests: Remove filename check
It is important that we check that the LaTeX bundle commands handle
filenames containing special characters correctly. This means we also
create test files with names that contain characters not legal on
Windows but perfectly fine on macOS. As a consequence of this Windows
users will not be able to clone this repository.

 #ignore